requirements.txt
pymongo[srv]
google-adk
litellm
python-dotenv
pandas
sentence-transformers
